Warning fixes: 	* actions.c (id_cmp): Do pointer arithmetic as `long' not `int' to  	ensure enough bits for calculation. 	* ch-tree.h (check_text_length): Remove unused parameter. 	* convert.c (display_int_cst): Cast a HOST_WIDE_INT argument to 	function sprintf into the appropriate type for printing. 	* decl.c (print_lang_decl): Use HOST_WIDE_INT_PRINT_DEC as the 	format specifier. 	(print_mode): Likewise. 	(init_decl_processing): Cast the arguments of bcopy/bzero to char *. 	* grant.c (grant_array_type): Use HOST_WIDE_INT_PRINT_DEC as 	the format specifier. 	* inout.c (check_text_length): Remove unused parameter `type'. 	(build_chill_associate): Initialize variables `arg1', `arg2', 	`arg3', `arg4' and `arg5'. 	(build_chill_modify): Likewise. 	(scanformcont): Change type of variable `curr' to `unsigned char'. 	* lex.c (maybe_downcase): Cast the argument of `tolower' to 	`unsigned char'. 	* satisfy.c (satisfy): Remove unused parameter in call to 	`check_text_length'. 	* tasking.c (generate_tasking_code_variable): Pass a HOST_WIDE_INT 	as a `long' in call to function `error'. 	(decl_tasking_code_variable): Likewise.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@22679 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/ch/tasking.c
@@ -154,8 +154,8 @@ generate_tasking_code_variable (name, tasking_code_ptr, quasi_flag)
       /* check for value should be assigned is out of range */
       if (TREE_INT_CST_LOW (*tasking_code_ptr) >
 	  TREE_INT_CST_LOW (TYPE_MAX_VALUE (chill_taskingcode_type_node)))
-	  error ("Tasking code %d out of range for `%s'.",
-		 TREE_INT_CST_LOW (*tasking_code_ptr),
+	  error ("Tasking code %ld out of range for `%s'.",
+		 (long) TREE_INT_CST_LOW (*tasking_code_ptr),
 		 IDENTIFIER_POINTER (name));
     }
 
@@ -201,8 +201,8 @@ decl_tasking_code_variable (name, tasking_code_ptr, quasi_flag)
       /* check for value should be assigned is out of range */
       if (TREE_INT_CST_LOW (*tasking_code_ptr) >
 	  TREE_INT_CST_LOW (TYPE_MAX_VALUE (chill_taskingcode_type_node)))
-	  error ("Tasking code %d out of range for `%s'.",
-		 TREE_INT_CST_LOW (*tasking_code_ptr),
+	  error ("Tasking code %ld out of range for `%s'.",
+		 (long) TREE_INT_CST_LOW (*tasking_code_ptr),
 		 IDENTIFIER_POINTER (name));
     }
